backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
cors==1.0.1
cryptography==46.0.5
defusedxml==0.7.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.135.1
filelock==3.25.1
fonttools==4.62.0
fpdf2==2.8.7
future==1.0.0
gevent==25.9.1
greenlet==3.3.2
h11==0.16.0
httptools==0.7.1
idna==3.11
passlib==1.7.4
pillow==12.1.1
psycopg2-binary==2.9.11
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PySocks==1.7.1
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
PyYAML==6.0.3
requests==2.32.5
requests-file==3.0.1
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.48
starlette==0.52.1
tldextract==5.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
zope.event==6.1
zope.interface==8.2
